refactor: add navigation package to main
